+++ src/app/account-management/account-management-routing.module.ts
@@ -25,7 +25,7 @@ const routes: Routes = [
           component: DashboardComponent,
         },
         {
-          path: 'account',
+          path: 'business',
           component: AccountDataComponent,
         },
         {