+++ src/app/page-not-found/page-not-found.component.ts
@@ -11,5 +11,7 @@ export class PageNotFoundComponent implements OnInit {
 
   ngOnInit(): void {
   }
-
+  
 }
+
+